requirements.txt
django-braces
djangorestframework
django==1.11
requests==2.18.4
requests-cache==0.4.13
misaka==1.0.2
django-bootstrap3